Initial Commit - RJ::Utils
The purpose of this gem is to collect a set of utilities that I commonly
build and add to projects. Included are basic command pattern tools as
well as help for creating form objects and other useful abstractions to
decouple your business logic from Rails (or any other ruby framework).

Just used `bundle gem r_j-utils` to create this basic skeleton.